--- src/main/java/wooteco/subway/line/Line.java
@@ -1,48 +1,0 @@
-package wooteco.subway.line;
-
-public class Line {
-
-    private Long id;
-    private String name;
-    private String color;
-
-    public Line(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public Line(Long id, String name, String color) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public boolean isSameName(String name) {
-        return this.name.equals(name);
-    }
-
-    public boolean isSameId(Long id) {
-        return this.id.equals(id);
-    }
-
-    public void changeInfo(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public boolean isNotSameId(Long id) {
-        return !this.id.equals(id);
-    }
-}
