--- src/main/java/wooteco/subway/line/web/LineResponse.java
@@ -1,27 +1,0 @@
-package wooteco.subway.line.web;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import wooteco.subway.domain.Line;
-import wooteco.subway.station.StationResponse;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class LineResponse {
-    private Long id;
-    private String name;
-    private String color;
-    private List<StationResponse> stations;
-
-    public static LineResponse create(Line line) {
-        List<StationResponse> stations = line.stations().stream()
-                .map(StationResponse::create)
-                .collect(Collectors.toList());
-        return new LineResponse(line.getId(), line.getName(), line.getColor(), stations);
-    }
-}
