--- src/main/java/wooteco/subway/section/web/SectionRequest.java
@@ -1,31 +1,0 @@
-package wooteco.subway.section.web;
-
-import javax.validation.constraints.Positive;
-
-public class SectionRequest {
-    private Long upStationId;
-    private Long downStationId;
-    @Positive
-    private int distance;
-
-    public SectionRequest() {
-    }
-
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
