--- src/main/java/wooteco/subway/station/StationRequest.java
@@ -1,16 +1,0 @@
-package wooteco.subway.station;
-
-public class StationRequest {
-    private String name;
-
-    public StationRequest() {
-    }
-
-    public StationRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
