--- src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -1,18 +1,0 @@
-package wooteco.subway.station.dao;
-
-import wooteco.subway.domain.Station;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface StationDao {
-    Station save(Station station);
-
-    List<Station> findAll();
-
-    Optional<Station> findStationById(Long id);
-
-    Optional<Station> findStationByName(String name);
-
-    void delete(Long id);
-}
