+++ src/main/java/wooteco/subway/web/api/StationApiController.java
@@ -1,0 +1,44 @@
+package wooteco.subway.web.api;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import wooteco.subway.domain.Station;
+import wooteco.subway.service.StationService;
+import wooteco.subway.web.request.StationRequest;
+import wooteco.subway.web.response.StationResponse;
+
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RequiredArgsConstructor
+@RestController
+public class StationApiController {
+
+    private final StationService stationService;
+
+    @PostMapping("/stations")
+    public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
+        Station station = Station.from(stationRequest.getName());
+        Station newStation = stationService.save(station);
+        StationResponse stationResponse = new StationResponse(newStation.getId(), newStation.getName());
+        return ResponseEntity.created(URI.create("/stations/" + newStation.getId())).body(stationResponse);
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<StationResponse>> showStations() {
+        List<Station> stations = stationService.findAll();
+        List<StationResponse> stationResponses = stations.stream()
+                .map(it -> new StationResponse(it.getId(), it.getName()))
+                .collect(Collectors.toList());
+        return ResponseEntity.ok().body(stationResponses);
+    }
+
+    @DeleteMapping("/stations/{id}")
+    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
+        stationService.deleteStation(id);
+        return ResponseEntity.noContent().build();
+    }
+}
