+++ src/main/java/wooteco/subway/web/request/LineRequest.java
@@ -1,0 +1,49 @@
+package wooteco.subway.web.request;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Positive;
+
+public class LineRequest {
+    @NotBlank
+    private String name;
+    @NotBlank
+    private String color;
+    @NotNull
+    private Long upStationId;
+    @NotNull
+    private Long downStationId;
+    @Positive
+    private int distance;
+
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+        this.name = name;
+        this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
