+++ src/main/java/wooteco/subway/web/request/SectionRequest.java
@@ -1,0 +1,31 @@
+package wooteco.subway.web.request;
+
+import javax.validation.constraints.Positive;
+
+public class SectionRequest {
+    private Long upStationId;
+    private Long downStationId;
+    @Positive
+    private int distance;
+
+    public SectionRequest() {
+    }
+
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
