+++ src/main/java/wooteco/subway/web/request/StationRequest.java
@@ -1,0 +1,16 @@
+package wooteco.subway.web.request;
+
+public class StationRequest {
+    private String name;
+
+    public StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
