+++ src/main/java/wooteco/subway/web/response/LineResponse.java
@@ -1,0 +1,26 @@
+package wooteco.subway.web.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import wooteco.subway.domain.Line;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private List<StationResponse> stations;
+
+    public static LineResponse create(Line line) {
+        List<StationResponse> stations = line.stations().stream()
+                .map(StationResponse::create)
+                .collect(Collectors.toList());
+        return new LineResponse(line.getId(), line.getName(), line.getColor(), stations);
+    }
+}
